+++ otus-spring-hw3/src/main/java/ru/otus/spring/course/config/AppConfig.java
@@ -1,18 +1,10 @@
 package ru.otus.spring.course.config;
 
-import com.opencsv.CSVReader;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.MessageSource;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import ru.otus.spring.course.data.Line;
-import ru.otus.spring.course.service.AppService;
-import ru.otus.spring.course.service.ConsoleService;
-import ru.otus.spring.course.service.LocaleTextService;
-
-import java.io.InputStreamReader;
-import java.io.Reader;
-import java.nio.charset.StandardCharsets;
+import ru.otus.spring.course.service.*;
 import java.util.*;
 
 @Configuration
@@ -24,6 +16,9 @@ public class AppConfig {
     @Value("${app.infinite}")
     private String isInfinite;
 
+    @Value("${csv.file}")
+    private String csvFile;
+
     @Bean
     public ConsoleService consoleService() {
         return new ConsoleService(new Scanner(System.in), System.out);
@@ -31,31 +26,11 @@ public class AppConfig {
 
     @Bean
     public LocaleTextService localeTextService(MessageSource messageSource, ConsoleService consoleService) {
-        return new LocaleTextService(consoleService, language, messageSource);
+        return new LocaleTextServiceImpl(consoleService, language, messageSource);
     }
 
     @Bean
-    public List<Line> lines(@Value("${csv.file}") String csvFile) throws Exception {
-        Reader reader = new InputStreamReader(getClass().getResourceAsStream("/" + csvFile + "_" + language + ".csv"), StandardCharsets.UTF_8);
-        CSVReader csvReader = new CSVReader(reader);
-        List<Line> list = new ArrayList<>();
-        String[] nextRecord;
-        while ((nextRecord = csvReader.readNext()) != null) {
-            String question = nextRecord[0];
-            int answerIndex = Integer.valueOf(nextRecord[1]);
-            List<String> options = new ArrayList<>(Arrays.asList(nextRecord).subList(2, nextRecord.length));
-            list.add(Line.builder()
-                    .answer(answerIndex)
-                    .options(options)
-                    .question(question).build());
-        }
-        reader.close();
-        csvReader.close();
-        return list;
-    }
-
-    @Bean
-    public AppService appService(LocaleTextService localeTextService, List<Line> lines) {
-        return new AppService(localeTextService, lines, Boolean.parseBoolean(isInfinite));
+    public AppService appService(LocaleTextService localeTextService) {
+        return new AppServiceImpl(localeTextService, Boolean.parseBoolean(isInfinite), "/" + csvFile + "_" + language + ".csv");
     }
 }
