+++ otus-spring-hw3/src/main/java/ru/otus/spring/course/data/User.java
@@ -8,8 +8,8 @@ import java.util.Objects;
 @AllArgsConstructor
 @Getter
 public class User {
-    private String name;
-    private String surname;
+    private final String name;
+    private final String surname;
 
     @Override
     public boolean equals(Object o) {
